pyproject.toml
[tool.poetry]
name = "chatcmd"
version = "0.1.0"
description = ""
authors = ["Gvard Windlass <gvard.windlass@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
asyncpg = "^0.28.0"
python-dotenv = "^1.0.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.19"}
alembic = "^1.11.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pexpect = "^4.8.0"
factory-boy = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
